+++ app/src/main/java/com/wmdd/errandz/hirerPastJobDescription/HirerPastJobDescriptionActivity.java
@@ -178,7 +178,7 @@ public class HirerPastJobDescriptionActivity extends AppCompatActivity implement
 
         if (user.getTotalRating() > 0) {
             taskerRatingTextView.setText(String.format("%.1f", user.getTotalRating()));
-            taskerRatingBar.setRating(user.getTotalRating() / user.getNumberOfReviews());
+            taskerRatingBar.setRating((float) (user.getTotalRating() / user.getNumberOfReviews()));
         } else {
             taskerRatingTextView.setText("No Rating");
             taskerRatingBar.setVisibility(View.GONE);
